+++ src/AllegroApi/Clients/OrderManagementClient.cs
@@ -145,7 +145,7 @@ public class OrderManagementClient
             null,
             cancellationToken);
         
-        return await response.Content.ReadAsByteArrayAsync(cancellationToken);
+        return await response.Content.ReadAsByteArrayAsync(cancellationToken).ConfigureAwait(false);
     }
 
     /// <summary>